Move Visibility and MemberType out from OutlinePane
